Change test setup
Now the suite is responsible to provide a proper cluster.Config for
configuring the cluster.Fixture. After the fixture is created & started
the user still needs to update the config to point to the new fixture, a
function is provided for that.

Everything else like deletion of the temporary config file,
configuration of the kubectl (on fuxtue update) and that like is handled
by the suite itself.