Add namespace handling mechanisms
With `createAndUseNamespace()` we will create a random name, use the
global `kubeCtl` to create the namespace and configure it in the
kubeconfig.

With `KubeCtl.WithNamespace( Namespace{ ... } )` one instance of the
kubectl can be configured to use a special namespace. This which will not
be persistet to the kubeconfig.

The Namespace struct holds two members:
  `Name`: Represents the name of the namespace to be used or created. If
          kept empty a random name will be generated.
  `AutoCreate`: Will create the namesapce if it does not already exist.
                If it already exists, that namespace will be used, no
                error or warning will be emitted.