csi: Remove package global vars
This commit removes global state in the variables
- `csi.csiDrivers`
- `csi.nim`
- `csi.PluginHandler`

The CSI plugin struct now holds the methods to implement a volume plugin
and a registration handler. The state formerly held in the package
global variables now has been moved into this struct, too.
Therefore, any instance of the CSI plugin now is completely indipendent
and those not share anything with any other instance.

Also the function `csi.PluginToPluginHandler` is introduced, which
allows external components (e.g. the kubelet) to convert a volume plugin
to a plugin handler, given it is CSI plugin.

However:
The kubelet still relies on the shared state in `csi.PluginHandler`.
Therefore, we still expose the created CSI plugin (via a call to
`ProbeVolumePlugins`) as `csi.PluginHandler` but guard that via a
`sync.Once`.