csi: csiPlugin now implements a PluginHandler
The node info mamanger, the drivers store and the methods from the
registration handler have moved onto the csiDriver struct. Therefore the
csiPlugin also implements the PluginHandler interface.
We now got rid of all the package globals.

Except:
To still support the same contract with the kubelet, we implemented a
singleton csiDriver and expose that via a global variable, the kubelet
knows about, again.
We plan to change the kubelet to get hold of the PluginHandler via name
as opposed to this global variable in an upcoming PR. Once this is done,
it is easy to remove the global state / the singleton here.