kubelet: discover CSI plugin via name
Previously, the kubelet used a global variable in the `csi` package to
find the registration handler for the CSI plugin. Now this discover
process uses the name of the plugin and does not rely on the global
variable in the `csi` package anymore.

Additionally, because the the kubelet fails to starti now if the CSI
plugin cannot be found, the CSI plugin now also needs to be registered
for kubemark's hollow kubelet.